PZTRAC-30 Generowanie miesięcznych podsumowań wydatków i przychodów

+++ Model/Wallet.cs
@@ -11,7 +11,7 @@ namespace _2023pz_trrepo.Model
 
         public long IconId { get; set; }
 
-        public int AccountBalance { get; set; }
+        public decimal AccountBalance { get; set; }
         public string UserId { get; set; }
         public ICollection<Income> incomes { get; } = new List<Income>();
         public ICollection<Expenditure> expenditures { get; } = new List<Expenditure>();
